requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-formtools==2.2
django-heroku==0.3.1
django-mailgun==0.9.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Paperboy==1.0.1
phonenumbers==8.12.13
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
